requirements.txt
openshift
kubernetes
PyYAML
pexpect
pytest
python-autologs
bitmath
urllib3
pytest-jira
python-bugzilla

# pytest-marker-bugzilla
git+git://github.com/rhevm-qe-automation/pytest-marker-bugzilla.git#egg=pytest-marker-bugzilla
